server: tighten share auth CLI semantics
